[docs]  Add OSF community backend
The OSF perceval backend was forked from the weblate backend, and
developed primarily by Open@RIT member @suhascv, with some contributions
from me.  It currently supports collecting a number of global statistics
about OSF, as well as activity logs and forks for specific projects.

This commit adds a link to the perceval-osf backend to the list of
community-maintained backends in the README.

Ammended from bfc5fda:  Added signoff

Signed-off-by: Emi Simpson <emi@alchemi.dev>