[backend] Set default category if none is provided
The default category is now set to the first category listed in the backend's
`CATEGORIES` list. This change is necessary to accommodate the removal of
backend-specific `fetch` methods, which previously handled category assignment.

Updated the tests to validate the default and specific category assignments,
and removed the obsolete `test_fetch_archive_needs_category`.

Signed-off-by: Venu Vardhan Reddy Tekula <vt2182@nyu.edu>