[git] Update references changed on the remote repo only
When repositories are synched, the backend needs to update
the local references to the ones set on the remote repository.
The backend processed all the references, including those
up to date. This made the synchronization very slow.

With this commit, only the references that were updated on
the remote repository will be modified locally. We have
improved the performance by 2. Repositories are in sync
two or more times than before.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>