[nntp] Disable NNTP backend for Python 3.13
nntplib is no longer part of the Python standard library.
It was removed in Python 3.13 after being deprecated in
Python 3.11 (PEP 594).

This commit raises an exception when trying to use
the NNTP backend and nntplib is not available.

It shows a deprecated warning when using a lower version of
Python.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>