[git] Fix filename parsing for files with brackets
Perceval misinterpreted file paths containing curly brackets in
the stats section of a log. It incorrectly detects them as renamed
or moved. This was causing that the output contains new files
that are not associated to any action.

This fix ensures that a file is only considered renamed if the file path
contains '=>', reducing false positives of moved files.
This also includes a regex pattern to identify when a file is renamed.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>